requirements.txt
django
djangorestframework
django-blog
djangorestframework-simplejwt
djangorestframework-simplejwt-token-blacklist
